+++ src/http.ts
@@ -68,7 +68,7 @@ export class HttpClient implements Client {
             try {
                 rowsPromise = stream.query(hranaStmt);
             } finally {
-                stream.close();
+                stream.closeGracefully();
             }
 
             return resultSetFromHrana(await rowsPromise);
@@ -80,6 +80,7 @@ export class HttpClient implements Client {
     async batch(stmts: Array<InStatement>, mode: TransactionMode = "deferred"): Promise<Array<ResultSet>> {
         try {
             const hranaStmts = stmts.map(stmtToHrana);
+            const version = await this.#client.getVersion();
 
             // Pipeline all operations, so `hrana.HttpClient` can open the stream, execute the batch and
             // close the stream in a single HTTP request.
@@ -91,10 +92,14 @@ export class HttpClient implements Client {
                 const sqlCache = new SqlCache(stream, sqlCacheCapacity);
                 sqlCache.apply(hranaStmts);
 
-                const batch = stream.batch();
-                resultsPromise = executeHranaBatch(mode, batch, hranaStmts);
+                // TODO: we do not use a cursor here, because it would cause three roundtrips:
+                // 1. pipeline request to store SQL texts
+                // 2. cursor request
+                // 3. pipeline request to close the stream
+                const batch = stream.batch(false);
+                resultsPromise = executeHranaBatch(mode, version, batch, hranaStmts);
             } finally {
-                stream.close();
+                stream.closeGracefully();
             }
 
             return await resultsPromise;
@@ -105,7 +110,8 @@ export class HttpClient implements Client {
 
     async transaction(mode: TransactionMode = "write"): Promise<HttpTransaction> {
         try {
-            return new HttpTransaction(this.#client.openStream(), mode);
+            const version = await this.#client.getVersion();
+            return new HttpTransaction(this.#client.openStream(), mode, version);
         } catch (e) {
             throw mapHranaError(e);
         }
@@ -120,7 +126,7 @@ export class HttpClient implements Client {
             try {
                 promise = stream.sequence(sql);
             } finally {
-                stream.close();
+                stream.closeGracefully();
             }
 
             await promise;
@@ -143,8 +149,8 @@ export class HttpTransaction extends HranaTransaction implements Transaction {
     #sqlCache: SqlCache;
 
     /** @private */
-    constructor(stream: hrana.HttpStream, mode: TransactionMode) {
-        super(mode);
+    constructor(stream: hrana.HttpStream, mode: TransactionMode, version: hrana.ProtocolVersion) {
+        super(mode, version);
         this.#stream = stream;
         this.#sqlCache = new SqlCache(stream, sqlCacheCapacity);
     }